feat: date to spinner outcomes
