+++ src/logging/utils.ts
@@ -1,5 +1,7 @@
 import chalk from 'chalk';
 import { Logging } from '../interfaces/logging/beta';
+import { AntelopeLogging } from '../common/config';
+import { Log } from '../interfaces/logging/beta/listener';
 
 const COLOR_FUNCTIONS: Record<string, (text: string) => string> = {
   red: chalk.red,
@@ -9,6 +11,10 @@ const COLOR_FUNCTIONS: Record<string, (text: string) => string> = {
   magenta: chalk.magenta,
   white: chalk.white,
 };
+
+// Terminal control sequences
+export const NEWLINE = '\n';
+export const OVERWRITE_CURRENT_LINE = '\r\x1b[K';
 
 // map created this way because Logging.Level is undefined on firsts calls
 export const getLevelInfo = (() => {
@@ -111,4 +117,125 @@ export function stripAnsi(str: string): string {
     .replace(colorEscapeSequenceRegex(), '')
     .replace(terminalTitleSequenceRegex(), '')
     .replace(miscTerminalSequenceRegex(), '');
+}
+
+/**
+ * Serializes a value for logging, handling objects, arrays, and other types appropriately
+ * @param value - The value to serialize
+ * @returns A string representation of the value
+ */
+export function serializeLogValue(value: any): string {
+  if (value === null) return 'null';
+  if (value === undefined) return 'undefined';
+  if (typeof value === 'string') return value;
+  if (typeof value === 'number' || typeof value === 'boolean') return String(value);
+  if (value instanceof Error) return `${value.name}: ${value.message}`;
+  if (value instanceof Date) return value.toISOString();
+
+  // For objects and arrays, use JSON.stringify with proper formatting
+  try {
+    return JSON.stringify(value, null, 2);
+  } catch {
+    // Fallback for circular references or other serialization issues
+    return String(value);
+  }
+}
+
+/**
+ * Formats a date according to the specified format string
+ * Supports common format patterns like yyyy-MM-dd HH:mm:ss
+ *
+ * @param date - The date to format
+ * @param format - The format string
+ * @returns The formatted date string
+ */
+export function formatDate(date: Date, format = 'yyyy-MM-dd HH:mm:ss'): string {
+  // Default to ISO format if no format provided
+  if (!format) {
+    return date.toISOString();
+  }
+
+  const padZero = (num: number, length = 2) => String(num).padStart(length, '0');
+
+  // Format replacements
+  const replacements: Record<string, string> = {
+    yyyy: String(date.getFullYear()),
+    MM: padZero(date.getMonth() + 1),
+    dd: padZero(date.getDate()),
+    HH: padZero(date.getHours()),
+    mm: padZero(date.getMinutes()),
+    ss: padZero(date.getSeconds()),
+    SSS: padZero(date.getMilliseconds(), 3),
+    yy: String(date.getFullYear()).slice(-2),
+    M: String(date.getMonth() + 1),
+    d: String(date.getDate()),
+    H: String(date.getHours()),
+    m: String(date.getMinutes()),
+    s: String(date.getSeconds()),
+  };
+
+  // Replace patterns in the format string
+  let result = format;
+  for (const [pattern, value] of Object.entries(replacements)) {
+    result = result.replace(pattern, value);
+  }
+
+  return result;
+}
+
+/**
+ * Formats a log message with right-aligned date
+ * @param logging - The logging configuration
+ * @param log - The log event
+ * @param module - Optional module name
+ * @param defaultDateFormat - Default date format to use
+ * @returns The formatted message with right-aligned date
+ */
+export function formatLogMessageWithRightAlignedDate(
+  logging: AntelopeLogging,
+  log: Log,
+  module?: string,
+  defaultDateFormat = 'yyyy-MM-dd HH:mm:ss',
+): string {
+  const levelInfo = getLevelInfo(log.levelId as Logging.Level);
+  const message = log.args.map((arg) => serializeLogValue(arg)).join(' ');
+
+  let messageWithLevel: string;
+  if (log.levelId === Logging.Level.NO_PREFIX.valueOf()) {
+    messageWithLevel = message;
+  } else {
+    const levelText = `[${levelInfo.name}]`;
+    const coloredLevel = getColoredText(levelText, levelInfo.color);
+    messageWithLevel = `${coloredLevel} ${message}`;
+  }
+
+  if (logging.moduleTracking.enabled && module) {
+    messageWithLevel = `(${module}) ${messageWithLevel}`;
+  }
+
+  const dateStr = formatDate(new Date(log.time), logging.dateFormat || defaultDateFormat);
+  const dateText = chalk.gray(`[${dateStr}]`);
+
+  if (!isTerminalOutput()) {
+    return `${dateText} ${messageWithLevel}`;
+  }
+
+  const terminalWidth = process.stdout.columns || 80;
+  const dateWidth = stripAnsi(dateText).length;
+  const minGap = 2;
+  const lines = messageWithLevel.split('\n');
+
+  const dateStart = terminalWidth - dateWidth;
+
+  return lines
+    .map((line, idx) => {
+      const plainLine = stripAnsi(line);
+      if (idx === lines.length - 1) {
+        const padding = Math.max(0, dateStart - plainLine.length - minGap);
+        return line + ' '.repeat(padding) + dateText;
+      } else {
+        return line;
+      }
+    })
+    .join('\n');
 }
